Having the run command creates a conflict
When running the app in VS and also running `swa start` having the `run`
command also tries running the Client app. So both VS and swa try
running the Client. It seems to be fine throwing an error and then using
the VS version, but seems better to just remove it here.